+++ OUDS/Core/Themes/OrangeBusinessTools/Sources/OrangeBusinessToolsTheme.swift
@@ -104,6 +104,12 @@ import SwiftUI
 ///     let orangeTheme = OrangeBusinessToolsTheme(fontFamily: "Helvetica Neue Arabic")
 /// ```
 ///
+/// ## Tokens versions
+///
+/// - Core OUDS version: 1.9.0
+/// - Core Orange version: 1.2.0
+/// - Brand Orange Business Tools version: 2.3.0
+///
 /// - Since: 0.17.0
 public final class OrangeBusinessToolsTheme: OUDSTheme, @unchecked Sendable {