+++ OUDS/Core/Themes/Sosh/Sources/SoshTheme.swift
@@ -51,6 +51,12 @@ import SwiftUI
 ///
 /// This theme cannot be customized more.
 ///
+/// ## Tokens versions
+///
+/// - Core OUDS version: 1.9.0
+/// - Core Sosh version: 1.2.0
+/// - Brand Sosh version: 2.3.0
+///
 /// - Since: 0.17.0
 public final class SoshTheme: OUDSTheme, @unchecked Sendable {